REF: separate method-pinning functions (#27811)
